TinyCompress: fix error reporting in tinycompress
BZ: 107006

The compress API are expected to return a negative value on an error
with errno indicating the type of error. But in some cases, this is
not happening. This can cause issue in the upper layers.

The fix is the oops function always returns -1 and the actual
error number is determined by the global variable errno.

Change-Id: I7d91001206016869719cd54f28515077ea528460
Signed-off-by: Lakshmi N Vinnakota <lakshmi.n.vinnakota@intel.com>